feat: add the same workaround for getting texts directly in slots to VerifyText as we already have for ClickText.
Basically slots are considered invisible by default, since they do not have offsets (which is one of the checks we do). Since Salesforce seems to commonly use the practice of having texts directly in slots, we will take parent element from slots that have direct texts and are under <a> tag.